pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flops-benchmarking"
version = "0.1.0"
description = "FLOps Benchmarking Framework"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "FLOps Team @ DaSHLab BITS Goa"}
]

dependencies = [
    "flwr[simulation]>=1.11.0",
    "flwr-datasets>=0.3.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "psutil>=5.9.0",
    "wandb>=0.16.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "timm>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]
gpu = [
    "pynvml>=11.5.0",
]
jetson = [
    "jetson-stats>=4.2.0",
]
deployment = [
    "ansible>=8.0.0",
    "paramiko>=3.0.0",
    "fabric>=3.0.0",
]

[project.scripts]
flops-run = "src.main:main"
flops-plot = "scripts.plot:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

